add/remove TicketPool.java
Add addTicket method
Add removeTicket methode
test those methods using vendor thread and customer thread

+++ src/Ticket.java
@@ -3,9 +3,9 @@ import java.math.BigDecimal;
 public class Ticket {
     private int ticketId;
     private String eventName;
-    private BigDecimal ticketPrice;
+    private double ticketPrice;
 
-    public Ticket(int ticketId, String eventName, BigDecimal ticketPrice){
+    public Ticket(int ticketId, String eventName, double ticketPrice){
         this.ticketId = ticketId;
         this.eventName = eventName;
         this.ticketPrice = ticketPrice;
@@ -27,11 +27,11 @@ public class Ticket {
         this.eventName = eventName;
     }
 
-    public BigDecimal getTicketPrice() {
+    public double getTicketPrice() {
         return ticketPrice;
     }
 
-    public void setTicketPrice(BigDecimal ticketPrice) {
+    public void setTicketPrice(double ticketPrice) {
         this.ticketPrice = ticketPrice;
     }
